biQueries(refactor): cambia nombre carpetas y ruta
