+++ src/app/app.module.ts
@@ -25,7 +25,7 @@ import { PacienteAppService } from './monitor-activaciones/services/pacienteApp.
 import { SendMessageCacheService } from './monitor-activaciones/services/sendMessageCache.service';
 import { WebhookLogComponent } from './webhook-log/webhook-log.component';
 import { WebhookLogService } from './services/webhook-log.service';
-import { BIModule } from './b-i/b-i.module';
+import { BIModule } from './bi-queries/b-i.module';
 
 @NgModule({
   declarations: [
