+++ src/app/bi-queries/b-i-routing.module.ts
@@ -1,10 +1,10 @@
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { RoutingGuard } from '../login/routing-guard';
-import { BIComponent } from './b-i/b-i.component';
+import { BIComponent } from './bi-queries/b-i.component';
 
 const routes = [
-    { path: 'b-i', component: BIComponent, canActivate: [RoutingGuard]}
+    { path: 'bi-queries', component: BIComponent, canActivate: [RoutingGuard] }
 ];
 
 @NgModule({
