+++ src/app/bi-queries/bi-queries/b-i.component.ts
