+++ src/app/bi-queries/filtros/filtros.bi.decorator.ts
