+++ src/app/bi-queries/filtros/valor-numerico/valor-numerico.component.ts
@@ -12,9 +12,5 @@ import { filtrosBi } from '../filtros.bi.decorator';
 export class ValorNumericoComponent extends FiltroBiComponent implements OnInit {
 
   ngOnInit() {
-    // console.log(this.argumento);
-  }
-  emitChange(event) {
-    console.log(event);
   }
 }
