+++ src/app/bi-queries/interfaces/IArgumentoBi.interface.ts
@@ -6,7 +6,7 @@ export interface IArgumentoBi {
     params: string; // tal vez cambiar nombre por label
     componente: string;
     tipo: string;
-    nombre?: string; // name en html
+    nombre: string; // name en html
     valor: any;  // contenido del parámetro
-    required?: boolean; // campo requerido
+    required?: boolean; // campo requerido (aun no usado desde BD)
 }
