+++ src/app/bi-queries/interfaces/IFiltroBi.interface.ts
